requirements.txt
requests
transformers
torch
gtts
pydub
schedule
newspaper3k
flask
nltk